requirements.txt
fastapi
uvicorn
python-dotenv
httpx
twilio
sqlalchemy
sentence-transformers
faiss-cpu
openai
python-multipart
pydantic-settings
groq
pyngrok
